Improve feedback when syntax causes module to fail
When a module fails to execute the exception is modified to include the
module causing the problem. If it is a syntax error then the file is
parsed in js and the syntax problem is reported (if found). At the
moment the feedback is given via console.log so it is not treated as a
critical failure.